switched the data types to ints
for gauge and length i changed the data types to int so that it is easier to serialize and deserialize with the embedded device

+++ iOS/CoilWind/CoilWind/Model/CoilCoreDataClass.swift
@@ -16,15 +16,15 @@ public class Coil: NSManagedObject {
     struct coil:Encodable {
         var name: String?
         var color: String?
-        var gauge: Double
-        var length: Double
+        var gauge: Int
+        var length: Int
         var quantity: Int
 
     }
     
     struct sendCoil:Encodable{
-        var gauge: Double
-        var length: Double
+        var gauge: Int
+        var length: Int
         var quantity: Int
     }
 }
